# docs/swagger.yaml
basePath: /
definitions:
  internal_quotation.CreateQuotationRequest:
    properties:
      author:
        example: Sun Tzu
        type: string
      quote:
        example: Appear weak when you are strong, and strong when you are weak.
        type: string
    type: object
  internal_quotation.CreateQuotationResponse:
    properties:
      author:
        example: Sun Tzu
        type: string
      id:
        example: 1
        type: integer
      quote:
        example: Appear weak when you are strong, and strong when you are weak.
        type: string
    type: object
  internal_quotation.Quotation:
    properties:
      author:
        example: Sun Tzu
        type: string
      id:
        example: 1
        type: integer
      quote:
        example: Appear weak when you are strong, and strong when you are weak.
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: voronkovworkemail@gmail.com
    name: Andrew Voronkov
  description: API для работы с цитатами. Позволяет создавать, получать, удалять цитаты.
  title: Quotation Book API
  version: "1.0"
paths:
  /quotes:
    get:
      description: Возвращает список всех цитат
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_quotation.Quotation'
            type: array
      summary: Получить все цитаты
      tags:
      - quotes
    post:
      consumes:
      - application/json
      description: Добавляет новую цитату в список
      parameters:
      - description: Цитата для добавления
        in: body
        name: quote
        required: true
        schema:
          $ref: '#/definitions/internal_quotation.CreateQuotationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_quotation.CreateQuotationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать цитату
      tags:
      - quotes
  /quotes/{id}:
    delete:
      description: Удаляет цитату по ID
      parameters:
      - description: ID цитаты
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить цитату
      tags:
      - quotes
  /quotes/author/{author}:
    get:
      description: Возвращает все цитаты указанного автора
      parameters:
      - description: Автор цитаты
        in: path
        name: author
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_quotation.Quotation'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить цитаты по автору
      tags:
      - quotes
  /quotes/random:
    get:
      description: Возвращает случайную цитату из списка
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_quotation.Quotation'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить случайную цитату
      tags:
      - quotes
swagger: "2.0"